Moved NMS4ToNMSLegacy to convert1_to_legacy group of transformations
