+++ inference-engine/src/hetero_plugin/hetero_plugin.cpp
@@ -21,7 +21,6 @@
 #include <transformations/convert_opset1_to_legacy/convert_opset1_to_legacy.hpp>
 #include <transformations/convert_opset2_to_opset1/convert_opset2_to_opset1.hpp>
 #include <transformations/convert_opset3_to_opset2/convert_opset3_to_opset2.hpp>
-#include <transformations/convert_opset4/convert_opset4.hpp>
 
 using namespace InferenceEngine;
 using namespace InferenceEngine::PluginConfigParams;
@@ -79,7 +78,6 @@ InferenceEngine::ExecutableNetworkInternal::Ptr Engine::LoadExeNetworkImpl(const
             auto function = clonedNetwork->getFunction();
             ::ngraph::op::GenericIE::DisableReshape noReshape(function);
             ::ngraph::pass::CommonOptimizations().run_on_function(function);
-            ::ngraph::pass::ConvertOpSet4().run_on_function(function);
             ::ngraph::pass::ConvertOpSet3ToOpSet2().run_on_function(function);
             ::ngraph::pass::ConvertOpSet2ToOpSet1().run_on_function(function);
             ::ngraph::pass::ConvertOpSet1ToLegacy().run_on_function(function);
