+++ inference-engine/src/mkldnn_plugin/mkldnn_plugin.cpp
@@ -21,7 +21,6 @@
 #include <transformations/convert_opset1_to_legacy/convert_opset1_to_legacy.hpp>
 #include <transformations/convert_opset2_to_opset1/convert_opset2_to_opset1.hpp>
 #include <transformations/convert_opset3_to_opset2/convert_opset3_to_opset2.hpp>
-#include <transformations/convert_opset4/convert_opset4.hpp>
 #include <transformations/rt_info/fused_names_attribute.hpp>
 #include <ngraph/opsets/opset1.hpp>
 #include <ngraph/opsets/opset2.hpp>
@@ -78,7 +77,6 @@ static void Transformation(ICNNNetwork::Ptr& clonedNetwork) {
 
     // Note: instead of running all Conversion Transformations you can make up your own transformation pipeline
     ngraph::pass::CommonOptimizations(transformations_callback).run_on_function(nGraphFunc);
-    ngraph::pass::ConvertOpSet4(transformations_callback).run_on_function(nGraphFunc);
     ngraph::pass::ConvertOpSet3ToOpSet2(transformations_callback).run_on_function(nGraphFunc);
     ngraph::pass::ConvertOpSet2ToOpSet1(transformations_callback).run_on_function(nGraphFunc);
     ngraph::pass::ConvertOpSet1ToLegacy(transformations_callback).run_on_function(nGraphFunc);
