+++ inference-engine/src/transformations/include/transformations/convert_opset1_to_legacy/convert_opset1_to_legacy_tbl.hpp
@@ -63,3 +63,4 @@ NGRAPH_PASS(ConvertGatherTreeToGatherTreeIE, ::ngraph::pass)
 NGRAPH_PASS(ConvertTopKToTopKIE, ::ngraph::pass)
 NGRAPH_PASS(ConvertNMSToNMSIE, ::ngraph::pass)
 NGRAPH_PASS(ConstantFolding, ::ngraph::pass)
+NGRAPH_PASS(ConvertNMS4ToLegacy, ::ngraph::pass)
