+++ inference-engine/src/vpu/graph_transformer/src/frontend/frontend.cpp
@@ -21,7 +21,6 @@
 
 #include <convert_function_to_cnn_network.hpp>
 #include <generic_ie.hpp>
-#include <transformations/convert_opset4/convert_opset4.hpp>
 #include <transformations/convert_opset3_to_opset2/convert_opset3_to_opset2.hpp>
 #include <transformations/convert_opset2_to_opset1/convert_opset2_to_opset1.hpp>
 #include <transformations/convert_opset1_to_legacy/convert_opset1_to_legacy.hpp>
@@ -375,7 +374,6 @@ ModelPtr FrontEnd::runCommonPasses(ie::ICNNNetwork& network, const UnsupportedLa
             // Disable shape inference (WA for generic operations)
             ngraph::op::GenericIE::DisableReshape noReshape(nGraphFunc);
 
-            ngraph::pass::ConvertOpSet4().run_on_function(nGraphFunc);
             ngraph::pass::ConvertOpSet3ToOpSet2().run_on_function(nGraphFunc);
             ngraph::pass::ConvertOpSet2ToOpSet1().run_on_function(nGraphFunc);
             ngraph::pass::ConvertOpSet1ToLegacy().run_on_function(nGraphFunc);
